Remove final Bento usage from project
