--- .bento/archive.json
@@ -1,51 +1,0 @@
-{
-  "bandit": {
-    "violations": {
-      "4e0db4785748aff9799910743970a07b": {
-        "tool_id": "bandit",
-        "check_id": "start-process-with-a-shell",
-        "path": "semgrep/setup.py",
-        "message": "Starting a process with a shell, possible injection detected, security issue.",
-        "severity": 2,
-        "syntactic_context": "os.system(  # nosem: contrib.dlint.dlint-equivalent.insecure-os-exec-use\n    os.path.join(repo_root, \"release-scripts/ubuntu-release.sh\")"
-      },
-      "54f98544761d0bae8b9ed0b9dc091c71": {
-        "tool_id": "bandit",
-        "check_id": "exec-used",
-        "path": "semgrep/semgrep/evaluation.py",
-        "message": "Use of exec detected.",
-        "severity": 1,
-        "syntactic_context": "exec(to_eval, scope)  # nosem: contrib.dlint.dlint-equivalent.insecure-exec-use"
-      },
-      "5894895d3058ff58d8c75da279c6eced": {
-        "tool_id": "bandit",
-        "check_id": "start-process-with-a-shell",
-        "path": "semgrep/setup.py",
-        "message": "Starting a process with a shell, possible injection detected, security issue.",
-        "severity": 2,
-        "syntactic_context": "os.system(  # nosem: contrib.dlint.dlint-equivalent.insecure-os-exec-use\n    os.path.join(repo_root, \"release-scripts/osx-release.sh\")"
-      }
-    }
-  },
-  "dlint": {
-    "violations": {}
-  },
-  "flake8": {
-    "violations": {}
-  },
-  "r2c.click": {
-    "violations": {}
-  },
-  "r2c.flask": {
-    "violations": {}
-  },
-  "r2c.registry.latest": {
-    "violations": {}
-  },
-  "r2c.requests": {
-    "violations": {}
-  },
-  "semgrep": {
-    "violations": {}
-  }
-}
