--- .bento/config.yml
@@ -1,35 +1,0 @@
-autorun:
-  block: true
-formatter:
-- clippy: {}
-tools:
-  bandit:
-    ignore:
-    - any-other-function-with-shell-equals-true
-    - assert-used
-    - hardcoded-bind-all-interfaces
-    - hardcoded-password-default
-    - hardcoded-password-funcarg
-    - hardcoded-password-string
-    - hardcoded-sql-expressions
-    - hardcoded-tmp-directory
-    - import-lxml
-    - import-pickle
-    - import-subprocess
-    - import-xml-expat
-    - import-xml-minidom
-    - import-xml-pulldom
-    - import-xml-sax
-    - import_xml-etree
-    - md5
-    - random
-    - start-process-with-no-shell
-    - start-process-with-partial-path
-    - subprocess-without-shell-equals-true
-    - try-except-continue
-    - urllib-urlopen
-    run: true
-  flake8:
-    ignore:
-    - bare-except-bugbear
-    run: true
